test(storefront): use migrate auto-dark-mode e2e | sas
