+++ packages/storefront/tests/e2e/specs/auto-dark-mode.e2e.ts
@@ -1,0 +1,30 @@
+import { getInternalUrls, getProperty } from '../helpers';
+import { TAG_NAMES } from '@porsche-design-system/shared';
+import { getComponentMeta } from '@porsche-design-system/component-meta';
+import { expect, test } from '@playwright/test';
+
+// filter out files from public/assets directory
+const internalUrls = getInternalUrls().filter((url) => !url.match(/^\/assets\/.*\.\w{3,4}$/));
+
+const themeableComponents = TAG_NAMES.filter((tagName) => getComponentMeta(tagName).isThemeable).join();
+const componentsWithThemeAutoSelector = `:where(${themeableComponents}):not(.playground *)`; // everything inside playground is not based on color-scheme preferences
+
+for (const [url, index] of internalUrls.map<[string, number]>((url, i) => [url, i])) {
+  test(`should have auto dark mode support at (${index + 1}/${internalUrls.length}) "${url}"`, async ({ page }) => {
+    await page.goto(url, { waitUntil: 'domcontentloaded' });
+    await page.waitForSelector('html.hydrated');
+    await page.evaluate(() =>
+      (window as unknown as Window & { componentsReady: () => Promise<number> }).componentsReady()
+    );
+
+    const components = await page.locator(componentsWithThemeAutoSelector);
+
+    // TODO: "The use of ElementHandle is discouraged, use Locator objects and web-first assertions instead.", see: https://playwright.dev/docs/api/class-elementhandle
+    for (const component of await components.elementHandles()) {
+      expect(
+        await getProperty(component, `theme`),
+        `"${await getProperty(component, 'tagName')}" didn't use theme="auto"`
+      ).toBe('auto');
+    }
+  });
+}
